requirements.txt
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.7.0
Brotli==1.1.0
certifi==2024.7.4
cffi==1.16.0
cryptography==42.0.8
exceptiongroup==1.2.1
greenlet==3.0.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.7
kaitaistruct==0.10
outcome==1.3.0.post0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycparser==2.22
pyOpenSSL==24.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dotenv==1.0.1
schedule==1.2.2
selenium==4.22.0
selenium-wire==5.1.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
SQLAlchemy==2.0.31
trio==0.26.0
trio-websocket==0.11.1
typing_extensions==4.12.2
urllib3==2.2.2
websocket-client==1.8.0
wsproto==1.2.0
zstandard==0.22.0
